If and conditionals done
There is a fair bit of magic in how the conditionals are implemented.
We use the fact that if considers anything non-zero to be true to our
advantage to keep the code short. We don't care if the value is 1 or
not, just that it's not zero. We also assume that memory location 000
is never 000 (usually true unless someone has self-modifying code, which
wouldn't be the case if they're using this compiler).